+++ src/utils/host.rs
@@ -1,5 +1,7 @@
 use std::path::PathBuf;
 
+// TODO: The flatpak versions are wrong
+
 pub fn user_data_dir() -> PathBuf {
     match std::env::var_os("HOST_XDG_DATA_HOME") {
         Some(dir) if ashpd::is_sandboxed() => dir.into(),
@@ -17,6 +19,6 @@ pub fn user_cache_dir() -> PathBuf {
 pub fn user_config_dir() -> PathBuf {
     match std::env::var_os("HOST_XDG_CONFIG_HOME") {
         Some(dir) if ashpd::is_sandboxed() => dir.into(),
-        _ => glib::user_cache_dir(),
+        _ => glib::user_config_dir(),
     }
 }